fix(u8): address thin client QA findings
